requirements.txt
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
astroid==3.3.5
asttokens==3.0.0
async-timeout==4.0.3
attrs==24.3.0
beautifulsoup4==4.12.3
bleach==6.1.0
brotlipy==0.7.0
captcha==0.4
certifi==2024.8.30
cffi==1.17.1
chardet
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
#conda==4.9.2
conda-package-handling
contourpy==1.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==43.0.1
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.8.0rc2
dill==0.3.9
Django==4.2.17
django-active-link==0.2.2
django-allauth==65.0.2
django-appconf==1.0.6
django-authtools==2.0.1
django-bootstrap5==24.3
django-braces==1.16.0
django-ckeditor==6.7.1
django-ckeditor-5==0.2.15
django-classy-tags==4.1.0
django-common-helpers==0.9.2
django-cookie-law==2.2.0
django-countries==7.6.1
django-crispy-forms==1.9.0
django-cron==0.6.0
django-crontab==0.7.1
django-debug-toolbar==4.3.0
django-environ==0.11.2
django-haystack==3.3.0
django-js-asset==2.2.0
django-leaflet==0.31.0
django-machina==1.3.1
django-modeltranslation==0.18.11
django-mptt==0.16.0
django-oauth-toolkit==2.4.0
django-oauth2-provider==0.2.6.1
django-oidc-provider==0.8.2
django-rated-reviews==1.0.1
django-ratelimit==4.1.0
django-recaptcha==4.0.0
django-redis==5.4.0
django-rest-swagger==2.2.0
django-rq==3.0.0
django-select2==8.2.1
django-summernote==0.8.20.0
django-templated-mail==1.1.1
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-spectacular==0.28.0
drf-yasg==1.21.8
easy-thumbnails==2.9
exceptiongroup==1.2.2
executing==2.1.0
fontawesomefree==6.6.0
fonttools==4.43.1
future==1.0.0
geographiclib==2.0
geopy==2.4.1
idna==3.10
importlib-resources==6.1.0
importlib_metadata==8.5.0
inflection==0.5.1
ipykernel==6.29.5
ipython==8.18.1
isort==5.13.2
itypes==1.2.0
jedi==0.19.2
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jwcrypto==1.5.6
kiwisolver==1.4.5
lazy-object-proxy==1.10.0
markdown2==2.5.1
MarkupSafe==3.0.2
matplotlib==3.8.1
matplotlib-inline==0.1.7
mccabe==0.7.0
nest-asyncio==1.6.0
numpy==1.26.1
oauthlib==3.2.2
openapi-codec==1.3.2
openrouteservice==2.3.3
packaging==24.1
pandas==2.1.2
parso==0.8.4
pep8==1.7.1
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.1
psycopg2==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.7.0
pycosat
pycparser==2.22
pycryptodomex==3.21.0
Pygments==2.19.1
pyjwkest==1.4.2
PyJWT==2.9.0
pylint==3.3.1
pyOpenSSL
pyparsing==3.1.1
PySocks
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
redis==5.1.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.22.3
rq==2.0.0
#ruamel-yaml-conda
#ruamel.yaml==0.18.6
#ruamel.yaml.clib==0.2.12
scipy
shortuuid==1.0.11
simplejson==3.19.3
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
soupsieve==2.6
sqlparse==0.5.1
stack-data==0.6.3
tomli==2.0.2
tomlkit==0.13.2
tornado==6.4.2
tqdm
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.2.3
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.4
Whoosh==2.7.4
wrapt==1.16.0
zipp==3.21.0
